Login system
Created a simple user login system with a registration page, a login page, and a logout link in the header.
There is a link in the header that leads to the registration page.
On the registration page (/register route), a username and password pair can be created that gets stored in the database.
Password storage and retrieval uses salted password hashing for maximum security.
If either field is empty while clicking the Submit button on the registration page, the Please, fill in both fields. error message is displayed.
If the username field contains a username that already exists in the database while clicking the Submit button on the registration page, the Username already exists, please choose another one! error message is displayed.
On successful registration, the Successful registration. Log in to continue. text is displayed, and the user is redirected to the login page.
There is a link in the header that leads to the login page.
On the login page (/login route), visitors can log in using the previously created username and password.
If the username and password pair does not match while clicking the Submit button on the login page, the Wrong username or password. error message is displayed.
Clicking the logout link (/logout route) logs the user out.